revert sample fn filter
