+++ examples/dpo_trainer/README.md
@@ -99,24 +99,6 @@ loss, so training does not load the SD3 VAE or text encoders during
 actor updates. Offline DPO also disables rollout and reward workers, so
 validation generation is disabled by default.
 
-### Sample Filtering
-
-You can reject weak win/lose pairs at dataloader sampling time. When a sampled
-row fails the filter, the dataset resamples until it finds a qualifying row.
-Configure a callable through `data.sample_filter_fn`:
-
-```bash
-bash examples/dpo_trainer/run_qwen_image_offline_dpo_lora.sh \
-  data.sample_filter_fn.path=pkg://verl_omni.utils.dataset.offline_dpo_dataset \
-  data.sample_filter_fn.name=offline_dpo_score_gap_filter \
-  data.sample_filter_fn.min_score_gap=0.07
-```
-
-The built-in `offline_dpo_score_gap_filter` keeps rows where
-`win_score - lose_score > min_score_gap` (default `0.07`). You can also point
-`path`/`name` to your own function that accepts a parquet row dict and returns
-`True` when the row should be kept.
-
 ## SD3.5 Training
 
 Stable Diffusion 3.5 is also supported. Use `--pipeline sd3` or leave
@@ -143,51 +125,6 @@ bash examples/dpo_trainer/run_sd35_medium_offline_dpo_lora.sh \
   data.train_files=data/offline_dpo_sd3/train.parquet \
   data.val_files=data/offline_dpo_sd3/test.parquet
 ```
-
-## Offline Checkpoint Validation
-
-After training, validate saved checkpoints on a prompt text file and plot
-training step vs. reward score:
-
-```bash
-python3 examples/dpo_trainer/validate_offline_dpo_checkpoints.py \
-  --pipeline qwen_image \
-  --model_path Qwen/Qwen-Image \
-  --prompt_file dataset/my_prompts/validation_prompts.txt \
-  --checkpoint_dir checkpoints/offline_dpo/qwen_image_offline_dpo_lora \
-  --output_dir outputs/offline_dpo_validation/qwen_image \
-  --reward_router_address 127.0.0.1:8000 \
-  --reward_model_name CodeGoat24/UnifiedReward-2.0-qwen3vl-8b \
-  --height 512 \
-  --width 512 \
-  --num_inference_steps 35 \
-  --true_cfg_scale 4.0
-```
-
-The script discovers `global_step_*/actor` checkpoints under
-`--checkpoint_dir`. You can also pass explicit checkpoint files or directories
-through `--checkpoint_paths`. For each checkpoint, it generates images for the
-validation prompts, calls the reward function, writes `validation_summary.jsonl`,
-`validation_details.jsonl`, CSV copies of both files, and saves
-`reward_curve.png`.
-
-When an actor directory contains multi-GPU FSDP shards such as
-`model_world_size_4_rank_*.pt`, the script merges them into a full checkpoint
-before loading. By default (`--fsdp_merge_backend auto`) it uses
-[`verl.model_merger`](https://verl.readthedocs.io/en/latest/advance/checkpoint.html)
-merge logic in-process and falls back to a local merge for diffusion LoRA
-checkpoints. To export a HuggingFace-format checkpoint manually, use:
-
-```bash
-python -m verl.model_merger merge \
-  --backend fsdp \
-  --local_dir checkpoints/offline_dpo/qwen_image_offline_dpo_lora/global_step_30/actor \
-  --target_dir /path/to/merged_hf_model \
-  --trust-remote-code
-```
-
-Or pass `--fsdp_merge_backend verl_cli` to run the same command inside the
-validation script (mainly for standard HuggingFace actor checkpoints).
 
 ## Reward Template
